Reapply theme components in a runnable.
We need to give other apps a chance to finish processing the
ACTION_PACKAGE_REPLACED before reapplying the updated theme
components.

An example of this behavior can be seen when a legacy icon pack is
applied and is updated.  Some Google services, such as google play,
FC when this happens.

Change-Id: If526223e6d318b4240fa765abb53dbe61f9ce746